+++ src/main/java/krasa/editorGroups/language/EditorGroupsCompletionContributor.kt
@@ -30,11 +30,7 @@ internal class EditorGroupsCompletionContributor : CompletionContributor() {
       CompletionType.BASIC,
       PlatformPatterns.psiElement(),
       object : CompletionProvider<CompletionParameters?>() {
-        override fun addCompletions(
-          parameters: CompletionParameters,
-          context: ProcessingContext,
-          result: CompletionResultSet
-        ) {
+        override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
           val file = parameters.originalFile
           val text = file.text