+++ src/main/java/krasa/editorGroups/tabs2/EditorGroupsTabsPosition.kt
@@ -3,4 +3,5 @@ package krasa.editorGroups.tabs2
 enum class EditorGroupsTabsPosition {
   TOP,
   BOTTOM,
+  LEFT,
 }